Untangle the CLI layer: typed runtime, shared wiring, split modules
- Replace the stringly-typed runtime dict with a frozen RuntimeOptions
  dataclass (api_key/api_base/fallback_model/pr_url) so the call-time
  contract is typed and self-documenting
- Extract build_provider_engine so the GitHub and local paths share one
  credential-resolution + provider-build path (also fixes the GitHub path
  ignoring cfg.api_base, which local mode already honored)
- Move output formatting to cli/render.py and Click command/option
  declarations to cli/commands.py; cli/__init__ keeps the wiring + execute
  logic (one namespace for the pipeline) and re-exports the public surface
