+++ src/lgtmaybe/cli/__init__.py
@@ -1,8 +1,14 @@
-"""CLI entry point for lgtmaybe.
+"""CLI for lgtmaybe.
 
-The command loads config (applying CLI precedence), resolves runtime options,
-then delegates to ``run_review`` which is injected with ports — testable with
-fakes, and wired with real adapters in the integration step.
+This package is split into three layers:
+
+- ``runtime`` / ``render`` — small, pure helpers (call-time options, output
+  formatting).
+- this module — the *logic*: parsing, adapter/provider wiring, and the
+  ``execute_*`` entry points that the commands call. Kept together so the
+  pipeline stages resolve (and can be patched in tests) as one namespace.
+- ``commands`` — the Click command + option declarations, imported at the
+  bottom to register onto the groups defined here.
 """
 
 from __future__ import annotations
@@ -10,14 +16,12 @@ from __future__ import annotations
 import json
 import os
 import re
-from pathlib import Path
 from typing import Any
 
 import click
 
-from lgtmaybe.cli.slash import dispatch, parse_command
-from lgtmaybe.config import store
-from lgtmaybe.config.loader import load_config
+from lgtmaybe.cli.render import render_findings
+from lgtmaybe.cli.runtime import RuntimeOptions
 from lgtmaybe.core.models import ReviewConfig, ReviewFinding
 from lgtmaybe.core.ports import GitHubGateway, ProviderClient, ReviewEngine
 from lgtmaybe.engine import LLMReviewEngine
@@ -43,66 +47,46 @@ def parse_pr_url(pr_url: str) -> tuple[str, int]:
     return f"{match['owner']}/{match['repo']}", int(match["number"])
 
 
-def render_findings(findings: list[ReviewFinding], summary: str, *, fmt: str = "human") -> str:
-    """Format findings for the local CLI.
+def build_provider_engine(
+    cfg: ReviewConfig, runtime: RuntimeOptions
+) -> tuple[LLMReviewEngine, ProviderClient]:
+    """Resolve credentials and build the provider + engine from config + runtime.
 
-    ``fmt`` selects the output: ``human`` (a readable listing + summary),
-    ``json`` (a machine-readable array), or ``agent`` (directive correction
-    instructions for an AI coding agent to read and apply).
+    Shared by every path that needs to talk to the model — the GitHub gateway
+    wiring and the local review alike — so credential resolution and provider
+    options stay in exactly one place. Raises ValueError with an actionable
+    message when a required credential is missing.
     """
-    if fmt == "json":
-        return json.dumps([f.model_dump(mode="json") for f in findings])
-    if fmt == "agent":
-        return _render_agent(findings, summary)
-
-    lines: list[str] = []
-    for f in findings:
-        lines.append(f"{f.path}:{f.line}  [{f.severity.upper()}] {f.title}")
-        lines.append(f"  {f.body}")
-        if f.suggestion is not None:
-            lines.append(f"  suggestion: {f.suggestion}")
-        lines.append("")
-    lines.append(summary)
-    return "\n".join(lines)
-
-
-def _render_agent(findings: list[ReviewFinding], summary: str) -> str:
-    """Render findings as correction instructions for an AI agent to apply."""
-    if not findings:
-        return f"No review findings — nothing to correct. {summary}"
-
-    lines = [
-        "Code review findings for your local changes. Act as the developer and "
-        "apply each correction below: open the file at the given path and line, "
-        "fix the issue, and apply the suggested change where one is given.",
-        "",
-    ]
-    for i, f in enumerate(findings, 1):
-        lines.append(f"[{i}] {f.path}:{f.line}  ({f.severity.upper()})  {f.title}")
-        lines.append(f"    Issue: {f.body}")
-        if f.suggestion is not None:
-            lines.append("    Suggested fix:")
-            lines.extend(f"        {s}" for s in f.suggestion.splitlines() or [f.suggestion])
-        lines.append("")
-    lines.append(
-        f"{len(findings)} finding(s) to address. After applying the fixes, re-run "
-        "`lgtmaybe review` to confirm they are resolved."
+    auth = resolve_credentials(
+        cfg.provider,
+        api_key=runtime.api_key,
+        api_base=runtime.api_base or cfg.api_base,
     )
-    return "\n".join(lines)
+    provider = build_provider(
+        cfg.provider,
+        cfg.model,
+        api_key=auth.api_key,
+        api_base=auth.api_base,
+        fallback_model=runtime.fallback_model,
+        timeout=cfg.timeout,
+        temperature=cfg.temperature,
+    )
+    return LLMReviewEngine(provider), provider
 
 
 def build_review_context(
-    cfg: ReviewConfig, runtime: dict[str, Any]
+    cfg: ReviewConfig, runtime: RuntimeOptions
 ) -> tuple[RestGitHubGateway, LLMReviewEngine, ProviderClient]:
     """Construct the gateway, engine, and provider from config + runtime.
 
-    Resolves provider credentials (ambient for cloud, key for the rest), builds a
-    litellm-backed provider, and points a REST gateway at the parsed PR. Raises
-    ValueError with an actionable message when a token or credential is missing.
-    The provider is returned too so slash commands (/ask, /describe) can use it
-    directly.
+    Builds the model side via ``build_provider_engine`` and points a REST gateway
+    at the parsed PR. Raises ValueError with an actionable message when the
+    GitHub token is missing. The provider is returned too so slash commands
+    (/ask, /describe) can use it directly.
     """
-    repo, pr_number = parse_pr_url(runtime["pr_url"])
+    if runtime.pr_url is None:
+        raise ValueError("a PR URL is required to build the GitHub review context")
+    repo, pr_number = parse_pr_url(runtime.pr_url)
 
     token = os.environ.get("GITHUB_TOKEN")
     if not token:
@@ -111,28 +95,13 @@ def build_review_context(
             "Set it in the environment (the GitHub Action provides it automatically)."
         )
 
-    auth = resolve_credentials(
-        cfg.provider,
-        api_key=runtime.get("api_key"),
-        api_base=runtime.get("api_base"),
-    )
-    provider = build_provider(
-        cfg.provider,
-        cfg.model,
-        api_key=auth.api_key,
-        api_base=auth.api_base,
-        fallback_model=runtime.get("fallback_model"),
-        timeout=cfg.timeout,
-        temperature=cfg.temperature,
-    )
-
+    engine, provider = build_provider_engine(cfg, runtime)
     github = RestGitHubGateway(repo=repo, pr_number=pr_number, token=token)
-    engine = LLMReviewEngine(provider)
     return github, engine, provider
 
 
 def build_adapters(
-    cfg: ReviewConfig, runtime: dict[str, Any]
+    cfg: ReviewConfig, runtime: RuntimeOptions
 ) -> tuple[GitHubGateway, ReviewEngine]:
     """Construct the real GitHub gateway and review engine from config + runtime."""
     github, engine, _provider = build_review_context(cfg, runtime)
@@ -160,142 +129,9 @@ def run_review(
     return findings, summary
 
 
-@click.group()
-def main() -> None:
-    """lgtmaybe — provider-agnostic PR reviewer."""
-
-
-@main.command()
-@click.option(
-    "--provider",
-    default=None,
-    help="LLM provider (openai, anthropic, bedrock, vertex, ollama, openrouter)",
-)
-@click.option("--model", default=None, help="Model name understood by the chosen provider")
-@click.option(
-    "--fallback-model",
-    default=None,
-    help="Model to retry with if the primary model fails",
-)
-@click.option(
-    "--api-key",
-    default=None,
-    envvar="LGTMAYBE_API_KEY",
-    help="API key (not needed for bedrock/vertex with ambient creds, or ollama)",
-)
-@click.option(
-    "--api-base", default=None, help="API base URL (useful for ollama: http://localhost:11434)"
-)
-@click.option(
-    "--min-severity",
-    default=None,
-    type=click.Choice(["info", "low", "medium", "high", "critical"]),
-    help="Minimum severity to report",
-)
-@click.option("--max-files", default=None, type=int, help="Maximum number of files to review")
-@click.option(
-    "--base",
-    default=None,
-    help="Base ref to diff the current branch against "
-    "(default: the remote's default branch, else main)",
-)
-@click.option(
-    "--working",
-    is_flag=True,
-    default=False,
-    help="Review uncommitted working-tree changes instead of the branch vs base",
-)
-@click.option(
-    "--format",
-    "output_format",
-    type=click.Choice(["human", "json", "agent"]),
-    default=None,
-    help="Output format: human listing (default), json array, or agent "
-    "(correction instructions for an AI coding agent to read and apply).",
-)
-@click.option(
-    "--json",
-    "as_json",
-    is_flag=True,
-    default=False,
-    help="Shorthand for --format json.",
-)
-@click.option(
-    "--context-lines",
-    default=None,
-    type=int,
-    help="Max unchanged lines added around each hunk for context (0 disables)",
-)
-@click.option(
-    "--timeout",
-    default=None,
-    type=int,
-    help="Per-request timeout in seconds for each model call (raise for slow local models)",
-)
-@click.option(
-    "--temperature",
-    default=None,
-    type=float,
-    help="Sampling temperature (default 0.0 for deterministic reviews)",
-)
-@click.option(
-    "--reflect/--no-reflect",
-    default=None,
-    help="Run the self-reflection pass that drops low-confidence findings "
-    "(--no-reflect keeps them all; useful for weaker models)",
-)
-@click.option(
-    "--config",
-    "config_path",
-    default=".lgtmaybe.yml",
-    show_default=True,
-    help="Path to a per-repo config file",
-)
-def review(
-    provider: str | None,
-    model: str | None,
-    fallback_model: str | None,
-    api_key: str | None,
-    api_base: str | None,
-    min_severity: str | None,
-    max_files: int | None,
-    base: str | None,
-    working: bool,
-    output_format: str | None,
-    as_json: bool,
-    context_lines: int | None,
-    timeout: int | None,
-    temperature: float | None,
-    reflect: bool | None,
-    config_path: str,
-) -> None:
-    """Review local git changes and print findings — no GitHub needed."""
-    cfg = load_config(
-        config_path=Path(config_path),
-        user_config_path=store.user_config_path(),
-        provider=provider,
-        model=model,
-        min_severity=min_severity,
-        max_files=max_files,
-        context_lines=context_lines,
-        timeout=timeout,
-        temperature=temperature,
-        reflect=reflect,
-    )
-
-    runtime: dict[str, Any] = {
-        "api_key": api_key,
-        "api_base": api_base,
-        "fallback_model": fallback_model,
-    }
-
-    fmt = output_format or ("json" if as_json else "human")
-    execute_local_review(cfg, runtime, base=base, working=working, fmt=fmt)
-
-
 def execute_local_review(
     cfg: ReviewConfig,
-    runtime: dict[str, Any],
+    runtime: RuntimeOptions,
     *,
     base: str | None,
     working: bool,
@@ -308,21 +144,7 @@ def execute_local_review(
     surfaces as a clean CLI error — there is no PR to post a notice to.
     """
     try:
-        auth = resolve_credentials(
-            cfg.provider,
-            api_key=runtime.get("api_key"),
-            api_base=runtime.get("api_base") or cfg.api_base,
-        )
-        provider = build_provider(
-            cfg.provider,
-            cfg.model,
-            api_key=auth.api_key,
-            api_base=auth.api_base,
-            fallback_model=runtime.get("fallback_model"),
-            timeout=cfg.timeout,
-            temperature=cfg.temperature,
-        )
-        engine = LLMReviewEngine(provider)
+        engine, _provider = build_provider_engine(cfg, runtime)
         ctx = local_pr_context(base=base, working=working)
         findings, summary = engine.review(ctx, cfg)
     except Exception as exc:
@@ -331,7 +153,7 @@ def execute_local_review(
     click.echo(render_findings(findings, summary, fmt=fmt))
 
 
-def execute_review(cfg: ReviewConfig, runtime: dict[str, Any], *, dry_run: bool) -> None:
+def execute_review(cfg: ReviewConfig, runtime: RuntimeOptions, *, dry_run: bool) -> None:
     """Build adapters, run the review, surface failures back to the PR.
 
     Shared by the ``review`` command and the ``action`` entrypoint.
@@ -357,111 +179,14 @@ def execute_review(cfg: ReviewConfig, runtime: dict[str, Any], *, dry_run: bool)
         click.echo(json.dumps([f.model_dump(mode="json") for f in findings]))
 
 
-def _post_failure(github: GitHubGateway, exc: Exception) -> None:
-    """Post a short failure notice to the PR; never raise from here."""
-    notice = f"⚠️ lgtmaybe review failed: {exc}"
-    try:
-        github.post_review([], notice)
-    except Exception:
-        # Posting the failure notice itself failed — nothing more we can do;
-        # the original error is still surfaced by the caller's ClickException.
-        pass
-
-
-@main.group(name="config")
-def config_cmd() -> None:
-    """Manage the user-level config (set provider/model/api_base once, reuse everywhere)."""
-
-
-@config_cmd.command("path")
-def config_path_command() -> None:
-    """Print the config file location."""
-    click.echo(str(store.user_config_path()))
-
-
-@config_cmd.command("show")
-def config_show() -> None:
-    """Print the current config."""
-    text = store.as_yaml()
-    click.echo(text if text else f"(no config yet at {store.user_config_path()})")
-
-
-@config_cmd.command("get")
-@click.argument("key")
-def config_get(key: str) -> None:
-    """Print one config value."""
-    value = store.get_key(key)
-    if value is not None:
-        click.echo(str(value))
-
-
-@config_cmd.command("set")
-@click.argument("key")
-@click.argument("value")
-def config_set(key: str, value: str) -> None:
-    """Set one config value (e.g. `config set model qwen3:27b`)."""
-    try:
-        coerced = store.set_key(key, value)
-    except ValueError as exc:
-        raise click.ClickException(str(exc)) from exc
-    click.echo(f"{key} = {coerced}")
-
-
-@config_cmd.command("init")
-def config_init() -> None:
-    """Interactively create the config file."""
-    provider = click.prompt("Provider", default="ollama")
-    model = click.prompt("Model", default="llama3")
-    api_base = click.prompt("API base (blank for none)", default="", show_default=False)
-    try:
-        store.set_key("provider", provider)
-        store.set_key("model", model)
-        if api_base.strip():
-            store.set_key("api_base", api_base)
-    except ValueError as exc:
-        raise click.ClickException(str(exc)) from exc
-    click.echo(f"Wrote {store.user_config_path()}")
-
-
-@main.command()
-@click.option(
-    "--event-path",
-    envvar="GITHUB_EVENT_PATH",
-    required=True,
-    help="Path to the issue_comment event payload (GitHub sets GITHUB_EVENT_PATH).",
-)
-@click.option("--provider", default=None, help="LLM provider override")
-@click.option("--model", default=None, help="Model name override")
-@click.option("--fallback-model", default=None, help="Model to retry with if the primary fails")
-@click.option("--api-key", default=None, envvar="LGTMAYBE_API_KEY", help="API key")
-@click.option("--api-base", default=None, help="API base URL (e.g. ollama)")
-@click.option("--config", "config_path", default=".lgtmaybe.yml", show_default=True)
-def comment(
-    event_path: str,
-    provider: str | None,
-    model: str | None,
-    fallback_model: str | None,
-    api_key: str | None,
-    api_base: str | None,
-    config_path: str,
-) -> None:
-    """Handle an issue_comment event: route a /slash command to the engine."""
-    event = json.loads(Path(event_path).read_text())
-    cfg = load_config(config_path=Path(config_path), provider=provider, model=model)
-    runtime: dict[str, Any] = {
-        "api_key": api_key,
-        "api_base": api_base,
-        "fallback_model": fallback_model,
-    }
-    execute_comment(event, cfg, runtime)
-
-
-def execute_comment(event: dict[str, Any], cfg: ReviewConfig, runtime: dict[str, Any]) -> None:
+def execute_comment(event: dict[str, Any], cfg: ReviewConfig, runtime: RuntimeOptions) -> None:
     """Route an issue_comment event's slash command to the engine/provider.
 
     Shared by the ``comment`` command and the ``action`` entrypoint. ``runtime``
     supplies api_key/api_base/fallback_model; the PR URL is derived here.
     """
+    from lgtmaybe.cli.slash import dispatch, parse_command
+
     parsed = parse_command(event.get("comment", {}).get("body", ""))
     if parsed is None:
         click.echo("No lgtmaybe slash command found; ignoring.")
@@ -474,7 +199,7 @@ def execute_comment(event: dict[str, Any], cfg: ReviewConfig, runtime: dict[str,
 
     repo = event["repository"]["full_name"]
     pr_number = issue["number"]
-    runtime = {**runtime, "pr_url": f"https://github.com/{repo}/pull/{pr_number}"}
+    runtime = runtime.with_pr_url(f"https://github.com/{repo}/pull/{pr_number}")
 
     try:
         github, engine, provider_client = build_review_context(cfg, runtime)
@@ -488,6 +213,17 @@ def execute_comment(event: dict[str, Any], cfg: ReviewConfig, runtime: dict[str,
         raise click.ClickException(f"/{parsed.name} failed: {exc}") from exc
 
 
+def _post_failure(github: GitHubGateway, exc: Exception) -> None:
+    """Post a short failure notice to the PR; never raise from here."""
+    notice = f"⚠️ lgtmaybe review failed: {exc}"
+    try:
+        github.post_review([], notice)
+    except Exception:
+        # Posting the failure notice itself failed — nothing more we can do;
+        # the original error is still surfaced by the caller's ClickException.
+        pass
+
+
 def pr_url_from_event(event: dict[str, Any]) -> str:
     """Build the PR URL from a pull_request(_target) event payload.
 
@@ -499,7 +235,7 @@ def pr_url_from_event(event: dict[str, Any]) -> str:
     return f"{server}/{repo}/pull/{number}"
 
 
-def _action_inputs() -> dict[str, str | None]:
+def action_inputs() -> dict[str, str | None]:
     """Read the action's declared inputs from the ``INPUT_*`` env vars.
 
     GitHub sets ``INPUT_<NAME>`` for each input of a container action; empty
@@ -519,31 +255,33 @@ def _action_inputs() -> dict[str, str | None]:
     }
 
 
-@main.command()
-def action() -> None:
-    """GitHub Action entrypoint: route by event, read inputs from env.
+@click.group()
+def main() -> None:
+    """lgtmaybe — provider-agnostic PR reviewer."""
 
-    ``issue_comment`` routes a slash command; any other event (``pull_request``
-    / ``pull_request_target``) runs a full review of the triggering PR.
-    """
-    inputs = _action_inputs()
-    cfg = load_config(
-        config_path=Path(inputs["config_path"] or ".lgtmaybe.yml"),
-        provider=inputs["provider"],
-        model=inputs["model"],
-    )
-    runtime: dict[str, Any] = {
-        "api_key": inputs["api_key"],
-        "api_base": None,
-        "fallback_model": inputs["fallback_model"],
-    }
 
-    event = json.loads(Path(os.environ["GITHUB_EVENT_PATH"]).read_text())
-    event_name = os.environ.get("GITHUB_EVENT_NAME", "")
+@main.group(name="config")
+def config_cmd() -> None:
+    """Manage the user-level config (set provider/model/api_base once, reuse everywhere)."""
 
-    if event_name == "issue_comment":
-        execute_comment(event, cfg, runtime)
-        return
 
-    runtime["pr_url"] = pr_url_from_event(event)
-    execute_review(cfg, runtime, dry_run=False)
+# Importing the commands module registers every command onto the groups above.
+# Done last so the logic functions the commands call are already defined.
+from lgtmaybe.cli import commands as _commands  # noqa: E402,F401
+
+__all__ = [
+    "RuntimeOptions",
+    "action_inputs",
+    "build_adapters",
+    "build_provider_engine",
+    "build_review_context",
+    "config_cmd",
+    "execute_comment",
+    "execute_local_review",
+    "execute_review",
+    "main",
+    "parse_pr_url",
+    "pr_url_from_event",
+    "render_findings",
+    "run_review",
+]
