+++ tests/cli/test_cli.py
@@ -7,7 +7,7 @@ import json
 import pytest
 from click.testing import CliRunner
 
-from lgtmaybe.cli import build_adapters, main, run_review
+from lgtmaybe.cli import RuntimeOptions, build_adapters, main, run_review
 from lgtmaybe.core.models import PRContext, ReviewConfig, ReviewFinding
 from lgtmaybe.core.ports import ReviewEngine
 from tests.fakes import FakeEngine, FakeGitHub, FakeProvider
@@ -167,7 +167,7 @@ class TestGitHubReviewErrorSurfacing:
         )
 
         with pytest.raises(click.ClickException):
-            cli_module.execute_review(_default_cfg(), {"pr_url": "x"}, dry_run=False)
+            cli_module.execute_review(_default_cfg(), RuntimeOptions(pr_url="x"), dry_run=False)
 
         assert len(github.posted) == 1
         posted_findings, posted_summary = github.posted[0]
@@ -223,11 +223,7 @@ class TestBuildAdapters:
 
         monkeypatch.setenv("GITHUB_TOKEN", "ghp_test")
         cfg = _default_cfg(provider="ollama", model="llama3")
-        runtime = {
-            "pr_url": "https://github.com/org/repo/pull/7",
-            "api_key": None,
-            "api_base": None,
-        }
+        runtime = RuntimeOptions(pr_url="https://github.com/org/repo/pull/7")
 
         github, engine = build_adapters(cfg, runtime)
 
@@ -237,11 +233,7 @@ class TestBuildAdapters:
     def test_requires_github_token(self, monkeypatch):
         monkeypatch.delenv("GITHUB_TOKEN", raising=False)
         cfg = _default_cfg(provider="ollama", model="llama3")
-        runtime = {
-            "pr_url": "https://github.com/org/repo/pull/7",
-            "api_key": None,
-            "api_base": None,
-        }
+        runtime = RuntimeOptions(pr_url="https://github.com/org/repo/pull/7")
 
         with pytest.raises(ValueError, match="GITHUB_TOKEN"):
             build_adapters(cfg, runtime)
@@ -251,11 +243,7 @@ class TestBuildAdapters:
         monkeypatch.setenv("GITHUB_TOKEN", "ghp_test")
         monkeypatch.delenv("OPENAI_API_KEY", raising=False)
         cfg = _default_cfg(provider="openai", model="gpt-4o")
-        runtime = {
-            "pr_url": "https://github.com/org/repo/pull/7",
-            "api_key": None,
-            "api_base": None,
-        }
+        runtime = RuntimeOptions(pr_url="https://github.com/org/repo/pull/7")
 
         with pytest.raises(ValueError, match="OPENAI_API_KEY"):
             build_adapters(cfg, runtime)
@@ -266,12 +254,9 @@ class TestBuildAdapters:
 
         monkeypatch.setenv("GITHUB_TOKEN", "ghp_test")
         cfg = _default_cfg(provider="ollama", model="llama3")
-        runtime = {
-            "pr_url": "https://github.com/org/repo/pull/7",
-            "api_key": None,
-            "api_base": None,
-            "fallback_model": "llama2",
-        }
+        runtime = RuntimeOptions(
+            pr_url="https://github.com/org/repo/pull/7", fallback_model="llama2"
+        )
 
         _github, _engine, provider = build_review_context(cfg, runtime)
